# Receive: Reuse strings in writer, scoped to a request