@matej-g 2023-01-31
This is superseded by https://github.com/thanos-io/thanos/pull/5926, closing.